requirements.txt
functions-framework==3.*
flask
Pillow>=10.0.0
google-cloud-storage>=2.0.0
